libhdb: add process debug options

+++ lib/hdb/process.cpp
@@ -11,8 +11,10 @@ module;
 export module hdb:process;
 
 import std;
-import :pipe;
+
+import :common;
 import :error;
+import :pipe;
 
 namespace hdb {
 
@@ -23,21 +25,22 @@ namespace hdb {
         };
 
         struct stop_reason {
-            stop_reason(int wait_status);
+            stop_reason(i32 wait_status);
 
             state reason;
-            int info;
+            u8 info;
         };
+
+        ~process();
 
         process() = delete;
         process(const process &) = delete;
         process &operator=(const process &) = delete;
 
-        friend std::unique_ptr< process > make_process(int pid, bool terminate_on_end);
-
-        ~process();
+        friend std::unique_ptr< process > make_process(i32 pid, bool terminate_on_end, bool debug);
 
         static std::unique_ptr< process > launch(std::filesystem::path path);
+        static std::unique_ptr< process > launch(std::filesystem::path path, bool debug);
         static std::unique_ptr< process > attach(pid_t pid);
 
         void resume();
@@ -48,12 +51,15 @@ namespace hdb {
         state state() const { return _state; }
       private:
 
-        process(pid_t pid, bool terminate_on_end)
-            : _pid(pid), _terminate_on_end(terminate_on_end)
+        process(pid_t pid, bool terminate_on_end, bool is_attached)
+            : _pid(pid)
+            , _terminate_on_end(terminate_on_end)
+            , _is_attached(is_attached)
         {}
 
-        pid_t _pid;
+        pid_t _pid = 0;
         bool _terminate_on_end = true;
+        bool _is_attached = true;
         enum state _state = state::stopped;
     };
 
@@ -62,35 +68,41 @@ namespace hdb {
             return;
         }
 
-        auto signal_and_wait = [this](int signal) {
+        auto signal_and_wait = [this](i32 signal) {
             kill(pid(), signal);
-            int status = 0;
+            i32 status = 0;
             waitpid(pid(), &status, 0);
         };
 
-        if (state() == state::running) {
-            signal_and_wait(SIGSTOP);
-        }
+        if (_is_attached) {
+            if (state() == state::running) {
+                signal_and_wait(SIGSTOP);
+            }
 
-        ptrace(PT_DETACH, pid(), {}, {});
-        kill(pid(), SIGCONT);
+            ptrace(PT_DETACH, pid(), {}, {});
+            kill(pid(), SIGCONT);
+        }
 
         if (_terminate_on_end) {
             signal_and_wait(SIGKILL);
         }
     }
 
-    std::unique_ptr< process > make_process(int pid, bool terminate_on_end) {
-        return std::unique_ptr< process >(new process(pid, terminate_on_end));
+    std::unique_ptr< process > make_process(i32 pid, bool terminate_on_end, bool debug) {
+        return std::unique_ptr< process >(new process(pid, terminate_on_end, debug));
     }
 
-    void exit_with_perror(hdb::pipe &channel, const std::string &prefix) {
+    [[noreturn]] void exit_with_perror(hdb::pipe &channel, const std::string &prefix) {
         auto msg = prefix + ": " + std::strerror(errno);
         channel.write(std::as_bytes(std::span(msg)));
         std::exit(EXIT_FAILURE);
     }
 
     std::unique_ptr< process > process::launch(std::filesystem::path path) {
+        return launch(path, /* debug */ true);
+    }
+
+    std::unique_ptr< process > process::launch(std::filesystem::path path, bool debug) {
         pipe channel(/* close on exec */ true);
 
         pid_t pid;
@@ -98,7 +110,7 @@ namespace hdb {
             error::send_errno("fork failed");
         } else if (pid == 0) {
             channel.close_read();
-            if (ptrace(PT_TRACE_ME, 0, {}, {}) < 0) {
+            if (debug && ptrace(PT_TRACE_ME, 0, {}, {}) < 0) {
                 exit_with_perror(channel, "tracing failed");
             }
 
@@ -116,8 +128,10 @@ namespace hdb {
             error::send(data);
         }
 
-        auto p = make_process(pid, /* terminate on end */ true);
-        p->wait_on_signal();
+        auto p = make_process(pid, /* terminate on end */ true, debug);
+        if (debug) {
+            p->wait_on_signal();
+        }
         return p;
     }
 
@@ -130,7 +144,7 @@ namespace hdb {
             error::send_errno("attach failed");
         }
 
-        auto p = make_process(pid, /* terminate on end */ false);
+        auto p = make_process(pid, /* terminate on end */ false, true);
         p->wait_on_signal();
         return p;
     }
@@ -143,24 +157,24 @@ namespace hdb {
         _state = state::running;
     }
 
-    process::stop_reason::stop_reason(int wait_status) {
+    process::stop_reason::stop_reason(i32 wait_status) {
         if (WIFSTOPPED(wait_status)) {
             reason = state::stopped;
-            info = WSTOPSIG(wait_status);
+            info = u8(WSTOPSIG(wait_status));
         } else if (WIFEXITED(wait_status)) {
             reason = state::exited;
-            info = WEXITSTATUS(wait_status);
+            info = u8(WEXITSTATUS(wait_status));
         } else if (WIFSIGNALED(wait_status)) {
             reason = state::terminated;
-            info = WTERMSIG(wait_status);
+            info = u8(WTERMSIG(wait_status));
         } else {
             error::send("unknown wait status");
         }
     }
 
     process::stop_reason process::wait_on_signal() {
-        int status = 0;
-        int options = 0;
+        i32 status = 0;
+        i32 options = 0;
         if (waitpid(pid(), &status, options) < 0) {
             error::send_errno("waitpid failed");
         }
